requirements.txt
openai>=1.40.0
pydantic>=2.7.0
PyYAML>=6.0.1
fastapi>=0.115.0
uvicorn>=0.30.0

psutil>=5.9.8
